Update everything to 'new' template
