+++ actions/issue-receiver/__tests__/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["../src", "**/*.test.ts"]
+}
