+++ actions/issue-receiver/package.json
@@ -5,9 +5,11 @@
   "main": "lib/main.js",
   "scripts": {
     "audit-moderate": "npm audit --audit-level=moderate",
+    "prebuild": "rimraf lib",
     "build": "tsc",
+    "prepackage": "npm run build && rimraf dist",
     "package": "ncc build",
-    "test": "jest"
+    "test": "jest --testRunner jest-circus/runner"
   },
   "repository": {
     "type": "git",
@@ -15,17 +17,27 @@
     "directory": "actions/issue-receiver"
   },
   "authors": [
-    "Sascha Mann <git@mail.saschamann.eu>"
+    "Sascha Mann <git@mail.saschamann.eu>",
+    "Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>"
   ],
-  "license": "AGPL-3.0",
+  "license": "AGPL-3.0-only",
   "dependencies": {
     "@actions/core": "^1.2.6",
     "@actions/github": "^4.0.0",
     "@octokit/webhooks": "^7.21.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-typescript": "^7.12.7",
+    "@types/jest": "^26.0.19",
+    "@types/nock": "^11.1.0",
     "@vercel/ncc": "^0.26.1",
+    "babel-jest": "^26.6.3",
     "jest": "^26.6.3",
+    "jest-circus": "^26.6.3",
+    "nock": "^13.0.5",
+    "rimraf": "^3.0.2",
     "typescript": "^4.1.3"
   }
 }
