+++ actions/notifier/action.yml
@@ -1,4 +1,4 @@
-name: Probby Notifier
+name: Probby push parser
 description: ''
 author: Sascha Mann <git@mail.saschamann.eu>
 
@@ -8,17 +8,14 @@ branding:
 
 inputs:
   token:
-    description: The PAT of the machine user.
-    required: true
+    description: The GH token used for accessing the API.
+    default: ${{ github.token }}
+    required: false
 
-  track-repos:
-    description: A comma-separated list of track repository to send a dispatch event to.
-    required: true
-
+outputs:
   payload-file:
     description: Path to the payload.json file that will be sent to the track repos.
-    required: true
 
 runs:
-  using: node12
-  main: dist/index.js
+  using: 'node12'
+  main: 'dist/index.js'
