+++ actions/notifier/src/main.ts
@@ -1,38 +1,6 @@
-import core from '@actions/core'
-import { context, getOctokit } from '@actions/github'
-import { readFileSync } from 'fs'
+import { run } from './run'
 
-async function run(): Promise<void> {
-    try {
-        // Init octokit client
-        const octokit = getOctokit(core.getInput('token'))
-
-        // Load dispatch payload
-        const payloadFile = core.getInput('payload-file')
-        const payload = JSON.parse(readFileSync(payloadFile).toString())
-
-        const repos = core.getInput('track-repos').split(',')
-
-        const events: Promise<unknown>[] = []
-
-        // Send dispatch event to repos
-        for (const repo of repos) {
-            core.debug(`Sending dispatch event to ${repo}...`)
-
-            events.push(
-                octokit.repos.createDispatchEvent({
-                    owner: context.repo.owner,
-                    repo: repo,
-                    event_type: 'probby-notification',
-                    client_payload: payload,
-                })
-            )
-        }
-
-        await Promise.all(events)
-    } catch (err) {
-        core.setFailed(err.message)
-    }
-}
-
+// The reason this is imported from a different file is that this is a
+// side-effect. That's fine when running this on the Github Action's Platform,
+// but not when doing anything else, such as running tests.
 run()
