+++ actions/parse-push/package.json
@@ -14,7 +14,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com:exercism/probby.git",
-    "directory": "actions/parse-pr"
+    "directory": "actions/parse-push"
   },
   "authors": [
     "Sascha Mann <git@mail.saschamann.eu>",
